backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
alembic==1.14.0
python-multipart==0.0.17
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.28.1
slowapi==0.1.9
redis==5.0.1


